feat: User 팔로워, 팔로잉 수 추가

+++ src/main/java/com/ssafy/nagne/service/FollowService.java
@@ -2,8 +2,14 @@ package com.ssafy.nagne.service;
 
 import static lombok.Lombok.checkNotNull;
 
+import com.ssafy.nagne.domain.User;
+import com.ssafy.nagne.error.DuplicatedFollowException;
+import com.ssafy.nagne.error.DuplicatedUnfollowException;
+import com.ssafy.nagne.error.NotFoundException;
 import com.ssafy.nagne.repository.FollowRepository;
+import com.ssafy.nagne.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -12,9 +18,10 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class FollowService {
 
+    private final UserRepository userRepository;
     private final FollowRepository followRepository;
 
-    public boolean save(Long id, Long followId) {
+    public boolean follow(Long id, Long followId) {
         checkNotNull(id, "id must be provided");
         checkNotNull(followId, "id must be provided");
 
@@ -22,7 +29,7 @@ public class FollowService {
             return false;
         }
 
-        return followRepository.save(id, followId) == 1;
+        return save(id, followId);
     }
 
     @Transactional(readOnly = true)
@@ -33,7 +40,7 @@ public class FollowService {
         return followRepository.check(id, followId);
     }
 
-    public boolean delete(Long id, Long followId) {
+    public boolean unfollow(Long id, Long followId) {
         checkNotNull(id, "id must be provided");
         checkNotNull(followId, "id must be provided");
 
@@ -41,6 +48,56 @@ public class FollowService {
             return false;
         }
 
+        return delete(id, followId);
+    }
+
+    private boolean save(Long id, Long followId) {
+        User me = findMe(id);
+        User follower = findFollower(followId);
+
+        follow(me, follower);
+
+        try {
+            return followRepository.save(id, followId) == 1;
+        } catch (DuplicateKeyException e) {
+            throw new DuplicatedFollowException("already followed");
+        }
+    }
+
+    private boolean delete(Long id, Long followId) {
+        User me = findMe(id);
+        User follower = findFollower(followId);
+
+        if (!check(id, followId)) {
+            throw new DuplicatedUnfollowException("already unfollowed");
+        }
+
+        unfollow(me, follower);
+
         return followRepository.delete(id, followId) == 1;
+    }
+
+    private User findMe(Long id) {
+        return userRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Could not found user for " + id));
+    }
+
+    private User findFollower(Long followId) {
+        return userRepository.findById(followId)
+                .orElseThrow(() -> new NotFoundException("Could not found user for " + followId));
+    }
+
+    private void follow(User me, User follower) {
+        me.follow(follower);
+
+        userRepository.update(me);
+        userRepository.update(follower);
+    }
+
+    private void unfollow(User me, User follower) {
+        me.unFollow(follower);
+
+        userRepository.update(me);
+        userRepository.update(follower);
     }
 }
